Update a comment and change int to double for a constant
Variable type changed so that there can be more precision

Comment changed to fit common Java conventions

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Core/TriMotorDrive.java
@@ -15,7 +15,7 @@ public class TriMotorDrive {
     //// CONSTANT VARIABLES
     private final int ENCODER_VALUES_PER_ROTATION = 1400;
     // TODO ADJUST VALUE
-    private final int INCHES_PER_ROTATION = 10;
+    private final double INCHES_PER_ROTATION = 10;
 
     /** Init */
     public TriMotorDrive (HardwareMap hardwareMap) {
@@ -50,7 +50,7 @@ public class TriMotorDrive {
     }
 
     /** Converts inches to encoder values using constants
-    * MAY BE UNRELIABLE, AS FRICTION IS UNACCOUNTED FOR */
+     * MAY BE UNRELIABLE, AS FRICTION IS UNACCOUNTED FOR */
     private int inchesToEncoderValues(double inches) {
         return Math.toIntExact(Math.round(inches * ENCODER_VALUES_PER_ROTATION / INCHES_PER_ROTATION));
     }
